pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "supply-chain-defender"
version = "0.1.0"
description = "Detect and respond to npm supply-chain attacks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
scd = "scd.cli:main"

[tool.setuptools.packages.find]
include = ["scd*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
